refactor(connectors): share the Postgres wire layer with Redshift
Extract the pg-protocol plumbing that Postgres and Redshift both need
into connectors/shared/pg-wire.ts: pool factory and pool types, pool
config resolution (parameterised by default port and connector label),
URL parsing, read-only query prep, and query/row helpers.

Postgres and Redshift now import this module instead of each carrying
their own copy. Each connector keeps its own metadata introspection
(Postgres pg_catalog, Redshift SVV_*), driver name, dialect, and
default port (5432 / 5439).

Postgres behaviour is unchanged: its test suite passes untouched, and
its public exports are preserved as thin aliases over the shared types.
Redshift picks up the statement_timeout / query-deadline handling it was
previously missing, matching Postgres.
