+++ packages/cli/src/connectors/redshift/connector.ts
@@ -1,5 +1,25 @@
-import { resolveStringReference } from '../shared/string-reference.js';
 import { getSqlDialectForDriver } from '../../context/connections/dialects.js';
+import { resolveQueryDeadlineMs, queryDeadlineExceededError } from '../../context/connections/query-deadline.js';
+import {
+  DefaultPgPoolFactory,
+  PG_OID_TYPE_MAP,
+  finiteNumber,
+  groupByTable,
+  isDeniedError,
+  isPgTimeoutError,
+  numberValue,
+  pgPoolConfigFromConfig,
+  preparePgReadOnlyQuery,
+  primaryKeyMap,
+  queryRows,
+  schemasFromConnection,
+  type KtxPgConnectionConfig,
+  type KtxPgEndpointResolver,
+  type KtxPgPool,
+  type KtxPgPoolConfig,
+  type KtxPgPoolFactory,
+  type KtxPgResolvedEndpoint,
+} from '../shared/pg-wire.js';
 import { assertReadOnlySql, limitSqlForExecution } from '../../context/connections/read-only-sql.js';
 import { tryConstraintQuery } from '../../context/scan/constraint-discovery.js';
 import { scopedTableNames } from '../../context/scan/table-ref.js';
@@ -26,95 +46,13 @@ import {
   type KtxTableSampleInput,
   type KtxTableSampleResult,
 } from '../../context/scan/types.js';
-import { Pool } from 'pg';
 
-const REDSHIFT_OID_TYPE_MAP: Record<number, string> = {
-  16: 'boolean',
-  20: 'bigint',
-  21: 'smallint',
-  23: 'integer',
-  25: 'text',
-  700: 'real',
-  701: 'double precision',
-  1043: 'varchar',
-  1082: 'date',
-  1114: 'timestamp',
-  1184: 'timestamptz',
-  1700: 'numeric',
-  2950: 'uuid',
-  3802: 'jsonb',
-  114: 'json',
-  1009: 'text[]',
-  1007: 'integer[]',
-  1016: 'bigint[]',
-};
-
-export interface KtxRedshiftConnectionConfig {
-  driver?: string;
-  host?: string;
-  port?: number;
-  database?: string;
-  username?: string;
-  user?: string;
-  password?: string;
-  url?: string;
-  schema?: string;
-  schemas?: string[];
-  ssl?: boolean;
-  sslmode?: string;
-  sslMode?: string;
-  rejectUnauthorized?: boolean;
-  maxConnections?: number;
-  [key: string]: unknown;
-}
-
-export interface KtxRedshiftPoolConfig {
-  host?: string;
-  port?: number;
-  database?: string;
-  user?: string;
-  password?: string;
-  connectionString?: string;
-  max: number;
-  idleTimeoutMillis: number;
-  connectionTimeoutMillis: number;
-  options?: string;
-  ssl?: { rejectUnauthorized: boolean };
-}
-
-interface KtxRedshiftQueryResult {
-  fields?: Array<{ name: string; dataTypeID: number }>;
-  rows: Record<string, unknown>[];
-}
-
-interface KtxRedshiftClient {
-  query(sql: string, params?: unknown[]): Promise<KtxRedshiftQueryResult>;
-  release(): void;
-}
-
-interface KtxRedshiftPool {
-  connect(): Promise<KtxRedshiftClient>;
-  end(): Promise<void>;
-  on?(event: 'error', listener: (error: Error) => void): void;
-}
-
-export interface KtxRedshiftPoolFactory {
-  createPool(config: KtxRedshiftPoolConfig): KtxRedshiftPool;
-}
-
-interface KtxRedshiftResolvedEndpoint {
-  host: string;
-  port: number;
-  close?: () => Promise<void>;
-}
-
-export interface KtxRedshiftEndpointResolver {
-  resolve(input: {
-    host: string;
-    port: number;
-    connection: KtxRedshiftConnectionConfig;
-  }): Promise<KtxRedshiftResolvedEndpoint>;
-}
+export type KtxRedshiftConnectionConfig = KtxPgConnectionConfig;
+export type KtxRedshiftPoolConfig = KtxPgPoolConfig;
+type KtxRedshiftPool = KtxPgPool;
+export type KtxRedshiftPoolFactory = KtxPgPoolFactory;
+type KtxRedshiftResolvedEndpoint = KtxPgResolvedEndpoint;
+export type KtxRedshiftEndpointResolver = KtxPgEndpointResolver;
 
 export interface KtxRedshiftScanConnectorOptions {
   connectionId: string;
@@ -203,133 +141,8 @@ interface RedshiftStatsRow {
   estimated_cardinality: unknown;
 }
 
-class DefaultRedshiftPoolFactory implements KtxRedshiftPoolFactory {
-  createPool(config: KtxRedshiftPoolConfig): KtxRedshiftPool {
-    return new Pool(config);
-  }
-}
-
-function groupByTable<T extends { table_name: string }>(rows: T[]): Map<string, T[]> {
-  const grouped = new Map<string, T[]>();
-  for (const row of rows) {
-    const tableRows = grouped.get(row.table_name) ?? [];
-    tableRows.push(row);
-    grouped.set(row.table_name, tableRows);
-  }
-  return grouped;
-}
-
 /** @internal */
-export function prepareRedshiftReadOnlyQuery(
-  sql: string,
-  params?: Record<string, unknown>,
-): { sql: string; params?: unknown[] } {
-  if (!params) {
-    return { sql, params: undefined };
-  }
-  const paramNames = Object.keys(params);
-  const values: unknown[] = new Array(paramNames.length);
-  const paramIndexMap = new Map<string, number>();
-  paramNames.forEach((name, index) => {
-    paramIndexMap.set(name, index + 1);
-    values[index] = params[name];
-  });
-  const sortedKeys = [...paramNames].sort((a, b) => b.length - a.length);
-  let parameterizedQuery = sql;
-  for (const name of sortedKeys) {
-    parameterizedQuery = parameterizedQuery.replace(new RegExp(`:${name}\\b`, 'g'), `$${paramIndexMap.get(name)}`);
-  }
-  return { sql: parameterizedQuery, params: values };
-}
-
-function primaryKeyMap(rows: RedshiftPrimaryKeyRow[]): Map<string, Set<string>> {
-  const grouped = new Map<string, Set<string>>();
-  for (const row of rows) {
-    const columns = grouped.get(row.table_name) ?? new Set<string>();
-    columns.add(row.column_name);
-    grouped.set(row.table_name, columns);
-  }
-  return grouped;
-}
-
-function isDeniedError(error: unknown): boolean {
-  if (!error || typeof error !== 'object') {
-    return false;
-  }
-  const code = (error as { code?: unknown }).code;
-  return code === '42501' || code === '42P01';
-}
-
-function queryRows(result: KtxRedshiftQueryResult): unknown[][] {
-  const headers = (result.fields ?? []).map((field) => field.name);
-  return result.rows.map((row) => headers.map((header) => row[header]));
-}
-
-function finiteNumber(value: unknown): number | null {
-  const parsed = Number(value);
-  return Number.isFinite(parsed) ? parsed : null;
-}
-
-function stringConfigValue(
-  connection: KtxRedshiftConnectionConfig | undefined,
-  key: keyof KtxRedshiftConnectionConfig,
-  env: NodeJS.ProcessEnv,
-): string | undefined {
-  const value = connection?.[key];
-  return typeof value === 'string' && value.trim().length > 0 ? resolveStringReference(value.trim(), env) : undefined;
-}
-
-
-function numberValue(value: unknown): number | undefined {
-  return typeof value === 'number' && Number.isFinite(value) ? value : undefined;
-}
-
-function positiveIntegerConfigValue(input: {
-  connection: KtxRedshiftConnectionConfig;
-  key: keyof KtxRedshiftConnectionConfig;
-  connectionId: string;
-  defaultValue: number;
-}): number {
-  const value = input.connection[input.key];
-  if (value === undefined) {
-    return input.defaultValue;
-  }
-  const numberValue = Number(value);
-  if (!Number.isInteger(numberValue) || numberValue < 1) {
-    throw new Error(`connections.${input.connectionId}.${String(input.key)} must be a positive integer`);
-  }
-  return numberValue;
-}
-
-function parseRedshiftUrl(url: string): Partial<KtxRedshiftConnectionConfig> {
-  const parsed = new URL(url);
-  const sslmode = parsed.searchParams.get('sslmode') ?? undefined;
-  return {
-    host: parsed.hostname,
-    port: parsed.port ? Number(parsed.port) : undefined,
-    database: parsed.pathname.replace(/^\/+/, '') || undefined,
-    username: parsed.username ? decodeURIComponent(parsed.username) : undefined,
-    password: parsed.password ? decodeURIComponent(parsed.password) : undefined,
-    ...(sslmode ? { sslmode } : {}),
-  };
-}
-
-function normalizedSslMode(connection: KtxRedshiftConnectionConfig): string | undefined {
-  const value = connection.sslmode ?? connection.sslMode;
-  return typeof value === 'string' && value.trim().length > 0 ? value.trim().toLowerCase() : undefined;
-}
-
-function schemasFromConnection(connection: KtxRedshiftConnectionConfig): string[] {
-  if (Array.isArray(connection.schemas) && connection.schemas.length > 0) {
-    return connection.schemas.filter((schema): schema is string => typeof schema === 'string' && schema.length > 0);
-  }
-  return typeof connection.schema === 'string' && connection.schema.length > 0 ? [connection.schema] : ['public'];
-}
-
-function searchPathSchemasFromConnection(connection: KtxRedshiftConnectionConfig): string[] {
-  const schemas = schemasFromConnection(connection);
-  return schemas.includes('public') ? schemas : [...schemas, 'public'];
-}
+export const prepareRedshiftReadOnlyQuery = preparePgReadOnlyQuery;
 
 export function isKtxRedshiftConnectionConfig(
   connection: KtxRedshiftConnectionConfig | undefined,
@@ -348,49 +161,13 @@ export function redshiftPoolConfigFromConfig(input: {
   if (!isKtxRedshiftConnectionConfig(input.connection)) {
     throw new Error(`Native Redshift connector cannot run driver "${inputDriver}"`);
   }
-
-  const env = input.env ?? process.env;
-  const referencedUrl = stringConfigValue(input.connection, 'url', env);
-  const urlConfig = referencedUrl ? parseRedshiftUrl(referencedUrl) : {};
-  const merged: KtxRedshiftConnectionConfig = { ...urlConfig, ...input.connection };
-  const host = stringConfigValue(merged, 'host', env);
-  const database = stringConfigValue(merged, 'database', env);
-  const user = stringConfigValue(merged, 'username', env) ?? stringConfigValue(merged, 'user', env);
-  const password = stringConfigValue(merged, 'password', env);
-  const sslmode = normalizedSslMode(merged);
-  const maxConnections = positiveIntegerConfigValue({
-    connection: merged,
-    key: 'maxConnections',
+  return pgPoolConfigFromConfig({
     connectionId: input.connectionId,
-    defaultValue: 10,
+    connection: input.connection,
+    defaultPort: 5439,
+    connectorLabel: 'Native Redshift connector',
+    env: input.env,
   });
-
-  if (!referencedUrl && !host) {
-    throw new Error(`Native Redshift connector requires connections.${input.connectionId}.host or url`);
-  }
-  if (!database && !referencedUrl) {
-    throw new Error(`Native Redshift connector requires connections.${input.connectionId}.database or url`);
-  }
-  if (!user && !referencedUrl) {
-    throw new Error(`Native Redshift connector requires connections.${input.connectionId}.username, user, or url`);
-  }
-
-  const config: KtxRedshiftPoolConfig = {
-    max: maxConnections,
-    idleTimeoutMillis: 30_000,
-    connectionTimeoutMillis: 10_000,
-    ...(referencedUrl && sslmode !== 'prefer' && sslmode !== 'disable'
-      ? { connectionString: referencedUrl }
-      : { host, port: numberValue(merged.port) ?? 5439, database, user, password }),
-  };
-  const searchPathSchemas = searchPathSchemasFromConnection(merged);
-  if (searchPathSchemas.length > 0) {
-    config.options = `-c search_path=${searchPathSchemas.join(',')}`;
-  }
-  if (merged.ssl && sslmode !== 'prefer' && sslmode !== 'disable') {
-    config.ssl = { rejectUnauthorized: merged.rejectUnauthorized ?? true };
-  }
-  return config;
 }
 
 export class KtxRedshiftScanConnector implements KtxScanConnector {
@@ -412,6 +189,7 @@ export class KtxRedshiftScanConnector implements KtxScanConnector {
   private readonly poolFactory: KtxRedshiftPoolFactory;
   private readonly endpointResolver?: KtxRedshiftEndpointResolver;
   private readonly now: () => Date;
+  private readonly deadlineMs: number;
   private readonly dialect = getSqlDialectForDriver('redshift');
   private pool: KtxRedshiftPool | null = null;
   private lastIdlePoolError: Error | null = null;
@@ -425,9 +203,10 @@ export class KtxRedshiftScanConnector implements KtxScanConnector {
       connection: options.connection,
       env: options.env,
     });
-    this.poolFactory = options.poolFactory ?? new DefaultRedshiftPoolFactory();
+    this.poolFactory = options.poolFactory ?? new DefaultPgPoolFactory();
     this.endpointResolver = options.endpointResolver;
     this.now = options.now ?? (() => new Date());
+    this.deadlineMs = resolveQueryDeadlineMs(this.connection);
     this.id = `redshift:${options.connectionId}`;
   }
 
@@ -807,11 +586,16 @@ export class KtxRedshiftScanConnector implements KtxScanConnector {
       const result = await client.query(assertReadOnlySql(sql), Array.isArray(params) ? params : undefined);
       return {
         headers: (result.fields ?? []).map((field) => field.name),
-        headerTypes: (result.fields ?? []).map((field) => REDSHIFT_OID_TYPE_MAP[field.dataTypeID] ?? `oid:${field.dataTypeID}`),
+        headerTypes: (result.fields ?? []).map((field) => PG_OID_TYPE_MAP[field.dataTypeID] ?? `oid:${field.dataTypeID}`),
         rows: queryRows(result),
         totalRows: result.rows.length,
         rowCount: result.rows.length,
       };
+    } catch (error) {
+      if (isPgTimeoutError(error)) {
+        throw queryDeadlineExceededError(this.deadlineMs, { cause: error });
+      }
+      throw error;
     } finally {
       client.release();
     }
